fix: move scale_bond_axis from algorithms to core to fix dependency inversion
core/mps.py was importing from algorithms/_tensor_utils.py, inverting the
intended layering (core should not depend on algorithms). Moved
scale_bond_axis and _scale_bond_axis_symmetric to a new core/_tensor_utils.py
module and kept a re-export in algorithms/_tensor_utils.py for backward
compatibility.

Closes #280
